requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
qrcode==7.3.1
reportlab==3.6.12
python-barcode==0.13.1
Pillow==9.5.0
python-dateutil==2.8.2